src/ol_openedx_course_translations/pyproject.toml
[project]
name = "ol-openedx-course-translations"
version = "0.2.0"
description = "An Open edX plugin to translate courses"
authors = [
  {name = "MIT Office of Digital Learning"}
]
license = "BSD-3-Clause"
readme = "README.rst"
requires-python = ">=3.11"
keywords = ["Python", "edx"]
dependencies = [
  "Django>=4.0",
  "djangorestframework>=3.14.0",
  "deepl>=1.25.0",
]

[project.entry-points."cms.djangoapp"]
ol_openedx_course_translations = "ol_openedx_course_translations.apps:OLOpenedXCourseTranslationsConfig"

[project.entry-points."lms.djangoapp"]
ol_openedx_course_translations = "ol_openedx_course_translations.apps:OLOpenedXCourseTranslationsConfig"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ol_openedx_course_translations"]
include = [
  "ol_openedx_course_translations/**/*.py",
]

[tool.hatch.build.targets.sdist]
include = [
  "ol_openedx_course_translations/**/*",
  "README.rst",
  "pyproject.toml",
]
